chessRush/requirements.txt
asgiref==3.5.2
Django==4.1.3
mysqlclient==2.1.1
sqlparse==0.4.3
tzdata==2022.7
django-rest-framework
djangorestframework-simplejwt
Pillow
python-barcode
django-ninja
drf-spectacular
django-cors-headers
whitenoise
chess
pandas